refactor(cli): move session usage module
